Added examples
The ace editors uses the Go syntax for now because it's similar enough.
But I'll have to create a proper ok plugin for ace later.

Some of the examples that rely on imports do not work yet because of the
way ok is built. I will have to fix that on the ok repo side.
