+++ app/.gitignore
@@ -1,0 +1,2 @@
+/packrd
+/main-packr.go
